Changes
- Fix compiler warnings
- Add operator to measure sizes of GOPs
- Fix tiling so that it doesn't add small tiles when offset = lastOffset.

+++ core/hevc/src/Stitcher.cc
@@ -74,8 +74,8 @@ namespace lightdb::hevc {
                 headerUpToPicOrder = current.headerUpToPicOrderCntLsb();
                 headerAfterPicOrder = current.headerAfterPicOrderCntLsb();
             } else {
-                assert(current.getEnd() == endLocationForP);
-                assert(current.getAddrOffset() == addrLocationForP);
+                assert(current.getEnd() == (unsigned int)endLocationForP);
+                assert(current.getAddrOffset() == (unsigned int)addrLocationForP);
                 assert(current.headerUpToPicOrderCntLsb() == headerUpToPicOrder);
                 assert(current.headerAfterPicOrderCntLsb() == headerAfterPicOrder);
             }
